Using Attributes to declare methods instead of relying on their names.
Validators and Handlers are operational.

PostHandler require the TResult to be handled:

The TResult is either:

- a IPoco and in which case it MUST be an IClosedPoco. This allows to
offer the same coherence through the modules as for the command
itself: the final command's  handler necessarily knows of all the "aspects"
of the result he has to produce. (Of course we check that the final
handler accepts the command with the TResult which is the Closure).

- another type (void, NoWaitResult, int, string, IChoucroute, etc.), in
which case all the TResults of a command must fit into a single
"specialization branch" and the "Closing Poco" of the command must be
of most specialized <TResult> .

This second option is only the "no Poco" version of the first one.
And its interest, its readability is questionable: we'll start with TResults
that are not IPoco can only be exactly of the same type.