Added Timeout config on HttpCrisSender.
A good API would be to support the timeout per call.
But for this, a cache of HttpClient wound be needed (using
a CancellationTokenSource/LinkToken is not an option because
a http timeout must be raised and not a CancelOperationException).

For the cache, one may use the IHttpClientFactory or roll our own
(the only instance variability is the timeout).