Some refactoring.
The creation of executed command is not ideal. When command is the
IAbstractCommand, the ExecutedCommand is untyped (it is not a
IExecutedCommand<T>, only a IExecutedCommand). To be rock solid
we need to introduce a typed factory of IExecutedCommand<T>:
the ICrisPocoModel is the place.

To avoid loose typing, a ICrisCommandModel : ICrisPocoModel should be
introduced.