New CK.Ng.Cris.AspNet.Auth package.
This tiny package updateAmbientValuesAsync when authentication
changes.

This is required because CK.TS.AspNet.Auth doesn't know HttpCrisEndpoint
and HttpCrisEndpoint (triggered by CK.Cris.TypeScript) can be used without
authentication.

The idea is that it is up to the "Application framework" to implement this
binding (here, for Angular, we use a APP_INITIALIZER provider but for
other framework this has to be done differently).