UserMessage transparent support!
CK.TypeScript now maps C# UserMessage type
to SimpleUserMessage TS type.
And the SimpleUserMessage can deserialize
the 8-cells array of a UserMessage in addition
to the regular 3-cells array [level,text,depth] of
the SimpleUserMessage. This guaranties that even
when the backend outputs Json UserMessages, the
TypeScript can read back its SimpleUserMessages.